Make distinction between provider and wallet
The providerOrSigner type becomes a bit clumsy in some circumstances so this
change favours either provider OR wallet, and leaves the providerOrSigner type
only to the contract calls.

Also, the term 'signer' is not used in Ethers.js so it can be confusing to use
since actually it is synonymous with 'wallet', so that term has been removed
completely.

The marketSide code has been adjusted so that, when implemented, either the
provider-connected version or the wallet-connected version needs to be chosen.
This helps narrow the definitions of the contract calls since some can be made
with just a provider while others need the more strict wallet connection.