Moved FloatChain.market to other file
The DX was really weird with this function:
```rescript
chain->FloatChain.market(marketIndex)
```
but now it's like this:
```
chain->FloatMarket.makeUsingChain(marketIndex)
```

Also did the same for MarketSide. Not there is a new file FloatMarketTypes which
is necessary to avoid a dependency cycle. I will improve the naming of these files
in a subsequent PR.